Fixed a bug where switching settings quickly would cause the LocationCheck to run both setting location checks.
settings command can now be ran in console

/game command can now be ran in console

Brawl items now has less arrows and has an anvil instead of ascension key

Added New Unlocks Obsidian, Daredevil

Added a upon crafting to Craft for events that are to happen upon craft

Added Boards into the Lobby

Fixed a bug where Entity Loot Tables where being rounded prematurely causing drops to have way worse chances than expected

Added a Info block to Unlocks GUI

** Balancing Changes **

Changed the Ascension Key Craft to Shapeless and requires two less diamonds and emeralds

Crimson Artifact is now black wool instead of black banner and now only costs 2 diamonds instead of 4

Changed Dread to require 4 less emeralds

Buffed Iron Pack and Gold Pack to be 12 ingots each

Notch Apple is now only 1 Craft (was 2)

Buffed Soul Artifact to be 3 Ghast Tears instead of 1

Changed the shape of Unlock A Weaver's Silk

Echolocator is now 3 emeralds 3 gold ingots was (6 1)

Ghost is now only 1 fermented Spider eye instead of 6 spider eyes and added 6 gold ingots

Pandora's box is now 4 Crafts instead of 3

Reduced the number of diamonds required in Red Ledger from 6 to 2

Modified Item Lists

+++ src/main/java/me/depickcator/ascension/LootTables/Entities/Entities/Bee.java
@@ -31,7 +31,7 @@ public class Bee implements LootTableChanger, EntityLootTable {
 
             Random r = new Random();
             int lootingLevel = getLootingLevel(e.getEntity().getKiller());
-            int honeycombCount = 1 + calculateLootingBonus(r, lootingLevel, 0, 1);
+            int honeycombCount = (int) (1 + calculateLootingBonus(r, lootingLevel, 0, 1));
             e.getEntity().getWorld().dropItem(e.getEntity().getLocation(), new ItemStack(Material.HONEYCOMB, honeycombCount));
         } catch (Exception ignored) {
             return false;
